test: move resource replay coverage into session shard
